Split classes into respective files
